requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
pyswisseph>=2.10.0
plotly>=5.15.0
scikit-learn>=1.3.0
catboost>=1.2.0
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
python-dotenv>=0.19.0
tqdm>=4.64.0